requirements/base.txt
django-cors-headers >= 3.14.0
django-environ==0.10.0
django-extensions==2.2.8
django-filter==22.1
django-phonenumber-field==7.0.2
django-url-checks==0.2.0
Django >= 4.1.7
djangorestframework >= 3.14.0
djangorestframework-simplejwt >= 5.2.2
djoser >= 2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
ipython==7.34.0
phonenumbers==8.13.8
pytz==2023.3
PyPDF2 >= 3.0.1
whitenoise==5.0.1
Pillow==9.5.0
